More refactoring + add docs
